+++ apps/backend/src/main.ts
@@ -9,14 +9,15 @@ import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 
 async function bootstrap() {
-  const app = await NestFactory.create(AppModule);
-  const globalPrefix = 'api';
-  app.setGlobalPrefix(globalPrefix);
-  const port = process.env.PORT || 3000;
-  await app.listen(port);
-  Logger.log(
-    `🚀 Application is running on: http://localhost:${port}/${globalPrefix}`
-  );
+    const app = await NestFactory.create(AppModule);
+    app.enableCors();
+    const globalPrefix = 'api';
+    app.setGlobalPrefix(globalPrefix);
+    const port = process.env.PORT || 3000;
+    await app.listen(port);
+    Logger.log(
+        `🚀 Application is running on: http://localhost:${port}/${globalPrefix}`,
+    );
 }
 
 bootstrap();
